Hybrid-SOC-Monitoring-Platform-Project/requirements.txt
opensearch-py==2.4.2
requests
python-dotenv
